% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipefittr.R
\name{pipefittr}
\alias{pipefittr}
\title{pipefittr}
\usage{
pipefittr(string, pretty = F)
}
\arguments{
\item{string}{a string, to be converted into \code{magrittr}'s pipe syntax}

\item{pretty}{create a multiline output, which is prettier. Try this.}
}
\description{
pipefittr
}

