% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilinepipefittr.r
\name{splitlisttodf}
\alias{splitlisttodf}
\title{Splits list into a data.frame}
\usage{
splitlisttodf(listtosplit)
}
\arguments{
\item{listtosplit}{a list to be converted into a data.frame}
}
\description{
Splits list into a data.frame
}

