% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilinepipefittr.r
\name{splitmultistrtolist}
\alias{splitmultistrtolist}
\title{Splits string into a list}
\usage{
splitmultistrtolist(stringtosplit)
}
\arguments{
\item{stringtosplit}{a string to be split}
}
\description{
Splits string into a list
}

