% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipefittr.R
\name{pipefittr}
\alias{pipefittr}
\title{Convert nested calls to magrittr's pipes.}
\usage{
pipefittr(string, pretty = F)
}
\arguments{
\item{string}{a string, to be converted into \code{magrittr}'s pipe syntax}

\item{pretty}{create a multiline output, which is prettier. Try this.}
}
\description{
To take nested function calls and convert them to a more readable form using magrittr's pipes.
}
\examples{
teststring = "jump_on(bop_on( scoop_up( hop_through(foo_foo, forest), field_mouse ), head))"
pipefittr(teststring, pretty = TRUE)
  

}

